+++ packages/std/src/encoding/decodeBase64Url.ts
@@ -7,5 +7,11 @@ import decodeBase64 from './decodeBase64';
  * decodeBase64Url('aGVsbG8') //=> Uint8Array [104, 101, 108, 108, 111]
  */
 export default function decodeBase64Url(str: string): Uint8Array {
+  if (/[+/]/.test(str)) {
+    throw new Error(
+      'Invalid Base64URL string: use only URL-safe characters.',
+    );
+  }
+
   return decodeBase64(str.replace(/-/g, '+').replace(/_/g, '/'));
 }